Refactor package layout and typo
