+++ src/main/java/com/codecool/shop/controller/ProductCategoryController.java
@@ -1,17 +1,10 @@
 package com.codecool.shop.controller;
 
-import com.codecool.shop.dao.implementation.ProductCategoryDaoMem;
-import com.codecool.shop.dao.implementation.ProductDaoMem;
-
-import com.codecool.shop.dao.implementation.SupplierDaoMem;
+import spark.ModelAndView;
 import spark.Request;
 import spark.Response;
-import spark.ModelAndView;
 
-import java.util.HashMap;
-import java.util.Map;
-
-public class ProductCategoryController extends Controller{
+public class ProductCategoryController extends ShopController {
 
     public static ModelAndView renderProductsByCategory(Request req, Response res) {
