+++ src/main/java/com/codecool/shop/controller/ProductController.java
@@ -1,19 +1,10 @@
 package com.codecool.shop.controller;
 
-import com.codecool.shop.dao.ProductCategoryDao;
-import com.codecool.shop.dao.ProductDao;
-import com.codecool.shop.dao.implementation.ProductCategoryDaoMem;
-import com.codecool.shop.dao.implementation.ProductDaoMem;
-import com.codecool.shop.dao.implementation.SupplierDaoMem;
-
+import spark.ModelAndView;
 import spark.Request;
 import spark.Response;
-import spark.ModelAndView;
 
-import java.util.HashMap;
-import java.util.Map;
-
-public class ProductController extends Controller{
+public class ProductController extends ShopController {
 
     public static ModelAndView renderProducts(Request req, Response res) {
