+++ src/main/java/com/codecool/shop/controller/SupplierController.java
@@ -1,17 +1,10 @@
 package com.codecool.shop.controller;
 
-import com.codecool.shop.dao.implementation.ProductCategoryDaoMem;
-import com.codecool.shop.dao.implementation.ProductDaoMem;
-import com.codecool.shop.dao.implementation.SupplierDaoMem;
-
+import spark.ModelAndView;
 import spark.Request;
 import spark.Response;
-import spark.ModelAndView;
 
-import java.util.HashMap;
-import java.util.Map;
-
-public class SupplierController extends Controller{
+public class SupplierController extends ShopController {
 
     public static ModelAndView renderProductsBySupplier(Request req, Response res) {
         return render(req, res);
